Dirty fix for duplicates in collection.
Seems like duplicates are caused by entity framework adding childs to the parent collection when each child is attached to context. Then AddElement method is adding the same child to the collection manually via reflection. This either related to proxies or foreign keys in models. Modified some tests trying to reproduce the issue, but so far no success.